Move code to middleware
